requirements.txt
streamlit~=1.40.1
pillow~=11.0.0
clearml
catboost
scikit-learn~=1.5.2
pandas~=2.2.3
numpy~=2.1.3
tqdm~=4.67.1
nltk~=3.9.1
scipy
optuna
joblib~=1.4.2
matplotlib~=3.9.3
Jinja2~=3.1.4
torch~=2.5.1
torchvision~=0.20.1
yadisk
clip
python-dotenv
shap
